# Coelsa/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Simulador COELSA — Cheques y Echeqs
  version: 1.0.0
  description: |
    Microservicio que simula el rol de COELSA (cámara electrónica de compensación)
    limitado a la gestión de cheques físicos (identificados por CMC7) y echeqs
    (identificados por IDECHEQ).

    **Convenciones clave:**
    - La creación es idempotente: enviar el header `Idempotency-Key` (GUID).
    - Las consultas por CUIT/CUIL están separadas por tipo y son paginadas.
    - Los errores se responden con ProblemDetails (RFC 7807) en español.
servers:
  - url: http://localhost:8080
    description: docker compose (local)
...
paths:
...
  /api/v1/cuentas:
    post:
      tags: [Cuentas]
      operationId: crearCuenta
      summary: Crea una cuenta corriente emisora de echeqs (idempotente)
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearCuentaRequest'
      responses:
        '201':
          description: Cuenta creada.
          headers:
            Idempotent-Replay:
              $ref: '#/components/headers/IdempotentReplay'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CuentaResponse'
        '200':
          description: Replay idempotente (misma key + mismo body).
          headers:
            Idempotent-Replay:
              $ref: '#/components/headers/IdempotentReplay'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CuentaResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    get:
      tags: [Cuentas]
      operationId: listarCuentas
      summary: Lista las cuentas activas de un CUIT/CUIL titular
      parameters:
        - $ref: '#/components/parameters/Cuit'
      responses:
        '200':
          description: Cuentas del titular.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CuentaResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/cuentas/{cbu}:
    get:
      tags: [Cuentas]
      operationId: obtenerCuenta
      summary: Obtiene una cuenta por su CBU
      parameters:
        - $ref: '#/components/parameters/Cbu'
      responses:
        '200':
          description: Cuenta encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CuentaResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/cuentas/{cbu}/chequeras:
    get:
      tags: [Cuentas]
      operationId: listarChequeras
      summary: Lista las chequeras de la cuenta, ordenadas por número
      parameters:
        - $ref: '#/components/parameters/Cbu'
      responses:
        '200':
          description: Chequeras de la cuenta.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChequeraResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Cuentas]
      operationId: solicitarChequera
      summary: Solicita una e-chequera nueva (50 números; reintentar con la misma key no duplica)
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Cbu'
      responses:
        '201':
          description: Chequera creada y vigente.
          headers:
            Idempotent-Replay:
              $ref: '#/components/headers/IdempotentReplay'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChequeraResponse'
        '200':
          description: Replay idempotente (misma key sobre el mismo CBU).
          headers:
            Idempotent-Replay:
              $ref: '#/components/headers/IdempotentReplay'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChequeraResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/cuentas/titulares:
    get:
      tags: [Cuentas]
      operationId: obtenerTitular
      summary: Padrón simulado ("lupa") — valida el documento y devuelve el nombre si tiene cuenta
      parameters:
        - name: tipoDoc
          in: query
          required: true
          description: CUIT, CUIL o CDI.
          schema:
            type: string
            enum: [CUIT, CUIL, CDI]
        - name: numero
          in: query
          required: true
          description: 11 dígitos con verificador.
          schema:
            type: string
            pattern: '^\d{11}$'
      responses:
        '200':
          description: Documento válido; nombre si está bancarizado en el simulador.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TitularResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: GUID único por operación de creación. Mismo key + mismo body = replay seguro.
      schema:
        type: string
        format: uuid
      example: 3f8a1c2e-9b4d-4f6a-8e2b-1d7c5a9e0f33
...
    Cuit:
      name: cuit
      in: query
      required: true
      description: CUIT/CUIL (librador o beneficiario). 11 dígitos con verificador módulo 11.
      schema:
        type: string
        pattern: '^\d{11}$'
      example: '20123456789'
    Cbu:
      name: cbu
      in: path
      required: true
      description: CBU de 22 dígitos con verificadores válidos (entidad + sucursal + cuenta).
      schema:
        type: string
        pattern: '^\d{22}$'
      example: '0110001300000000000017'
...
  headers:
    IdempotentReplay:
      description: "'true' si la respuesta es un replay idempotente de una creación previa."
      schema:
        type: boolean
        default: false
  responses:
    BadRequest:
      description: Error de validación (body inválido, CUIT inválido, falta Idempotency-Key, etc.).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Conflict:
      description: Conflicto — CBU/CMC7/IDECHEQ duplicado con key distinta, misma Idempotency-Key con body distinto, chequera sin números disponibles, tope de endosos/cesiones o pedido/cesión pendiente.
...
    NotFound:
      description: No existe un instrumento con el identificador indicado.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    TooManyRequests:
      description: Se excedió el límite de solicitudes (rate limiting, RF-08). Reintentar cuando indique Retry-After.
      headers:
        Retry-After:
          description: Segundos hasta que se pueda reintentar.
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  schemas:
    CrearCuentaRequest:
      type: object
      required: [cbu, cuitTitular, nombreTitular, moneda]
      properties:
        cbu:
          type: string
          pattern: '^\d{22}$'
          description: CBU con verificadores válidos.
          example: '0110001300000000000017'
        cuitTitular:
          $ref: '#/components/schemas/Cuit'
        nombreTitular:
          type: string
          maxLength: 120
          example: 'Alfa S.R.L.'
        moneda:
          $ref: '#/components/schemas/Moneda'
    CuentaResponse:
      type: object
      properties:
        cbu:
          type: string
          pattern: '^\d{22}$'
          description: El CBU de 22 dígitos (identificador de negocio).
          example: '0110001300000000000017'
        banco:
          type: string
          example: '011'
        sucursal:
          type: string
          example: '0001'
        numeroCuenta:
          type: string
          example: '0000000000001'
        cuitTitular:
          $ref: '#/components/schemas/Cuit'
        nombreTitular:
          type: string
          example: 'Alfa S.R.L.'
        moneda:
          $ref: '#/components/schemas/Moneda'
        fechaCreacion:
          type: string
          format: date-time
    ChequeraResponse:
      type: object
      properties:
        numero:
          type: integer
          format: int32
          description: Número de chequera (secuencial por cuenta).
          example: 1
        cantidadTotal:
          type: integer
          format: int32
          description: Números por chequera (50).
          example: 50
        proximoNumero:
          type: integer
          format: int32
          description: Próximo número a reservar.
          example: 8
        disponibles:
          type: integer
          format: int32
          example: 43
        estado:
          type: string
          enum: [Vigente, Agotada]
        fechaSolicitud:
          type: string
          format: date-time
        fechaHabilitacion:
          type: string
          format: date-time
    TitularResponse:
      type: object
      properties:
        tipoDoc:
          type: string
          enum: [CUIT, CUIL, CDI]
        numero:
          type: string
          pattern: '^\d{11}$'
        nombre:
          type: string
          nullable: true
          description: Nombre del titular si tiene cuenta; null si no está bancarizado.
          example: 'Alfa S.R.L.'
        bancarizado:
          type: boolean
...
    Moneda:
      type: string
      enum: [P, D]
      description: "'P' = pesos, 'D' = dólares."
...
    Cuit:
      type: string
      pattern: '^\d{11}$'
      description: CUIT/CUIL de 11 dígitos con dígito verificador (módulo 11).
      example: '20123456789'
    ProblemDetail:
      type: object
      description: Error RFC 7807 (application/problem+json), mensajes en español.
      properties:
        type:
          type: string
          example: 'https://tools.ietf.org/html/rfc7807'
        title:
          type: string
          example: 'Transición de estado inválida'
        status:
          type: integer
          example: 422
        detail:
          type: string
          example: 'No se puede pasar de Anulado a Compensado.'
        instance:
          type: string
          example: '/api/v1/cheques/060000114250000123400001234567/estado'
        errors:
          type: object
          nullable: true
          description: Diccionario campo → errores de validación.
          additionalProperties:
            type: array
            items:
              type: string
...